Add note about fencer name format in README
This commit updates the README file to add a note about the required format for the fencer name argument when running the script. The note reminds users to surround the name with quotes (e.g. 'Max Heinzer') to avoid errors caused by spaces or special characters. The usage example in the README has also been updated to reflect this change.